pyproject.toml
[tool.poetry]
name = "edubot"
version = "0.7.4"
description = ""
authors = ["exciteabletom <tom@digitalnook.net>", "moodler <martin@moodle.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = {version = "^2.0.14", extras = ["mypy"]}
openai = "^0.27.6"
stability-sdk = "^0.8.1"
pillow = "^9.5.0"
beautifulsoup4 = "^4.12.2"
trafilatura = "^1.6.0"
replicate = "^0.8.1"
tiktoken = "^0.4.0"
langchain = "^0.0.279"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.2"
mypy = "^1.3.0"

[tool.mypy]
plugins = "sqlalchemy.ext.mypy.plugin"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
